+++ src/main/java/com/aptus/blackbox/datamodels/Metering/EndpointMetering.java
@@ -3,7 +3,7 @@ package com.aptus.blackbox.datamodels.Metering;
 public class EndpointMetering {
 
 	private String endpoint;
-	private long totalRows;
+	private int totalRows;
 	
 	public String getEndpoint() {
 		return endpoint;
@@ -11,10 +11,10 @@ public class EndpointMetering {
 	public void setEndpoint(String endpoint) {
 		this.endpoint = endpoint;
 	}
-	public long getTotalRows() {
+	public int getTotalRows() {
 		return totalRows;
 	}
-	public void setTotalRows(long totalRows) {
+	public void setTotalRows(int totalRows) {
 		this.totalRows = totalRows;
 	}
